It produces some few errors on the json output(Extra commas)
The program causes some minor errors on the json output which is mainly insertion of commas where there is supposed to be none. After running the code several times the most number of errors produced on the json output is three. To be honest I have no idea why this is the case but am working on it. otherwise the program runs smoothly